refactor(cli): move to cobra
