+++ internal/buildinfo/buildinfo.go
@@ -1,0 +1,7 @@
+package buildinfo
+
+var (
+	Version = "dev"
+	Commit  = ""
+	Date    = ""
+)
