+++ main.go
@@ -1,0 +1,12 @@
+// Copyright (c) 2023 - 2024, nuxen and the seasonpackarr contributors.
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+package main
+
+import (
+	"seasonpackarr/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
